+++ app/Filament/Resources/BinomRumbleSpentDataResource.php
@@ -17,7 +17,7 @@ class BinomRumbleSpentDataResource extends Resource
 
     protected static ?string $navigationIcon = 'heroicon-o-banknotes';
     protected static ?string $navigationGroup = 'Rumble and Binom Reports Only';
-    protected static ?string $navigationLabel = '3. Binom Rumble Spent Datas';
+    protected static ?string $navigationLabel = '3. Binom Rumble Spent Data';
     protected static ?int $navigationSort = 3;
 
     public static function form(Form $form): Form