+++ app/Filament/Resources/RumbleDataResource.php
@@ -19,7 +19,7 @@ class RumbleDataResource extends Resource
 
     protected static ?string $navigationIcon = 'heroicon-o-rectangle-stack';
     protected static ?string $navigationGroup = 'Rumble and Binom Reports Only';
-    protected static ?string $navigationLabel = '1. Rumble Datas';
+    protected static ?string $navigationLabel = '1. Rumble Data';
     protected static ?int $navigationSort = 1;
 
     public static function form(Form $form): Form